Fail if item sets more than allowed values
Pretty much all records can be defaulted with an early / (which will
mostly likely be handled by a later stage), but it's reasonable to abort
already in the first phase if more-than-allowed entries are specified.

The bounded extension to item raises a parsing failure if it would
produce an item with more elements than expected.