Switch to AST generation
This patch changes the basic boost-spirit touching parser into
generating an AST, rather than trying to expand all wildcards, defaults
and enforce some constraints.

The main motivation for this is to not having to specify every keyword's
structure multiple times (because it's needed for semantic analysis),
and there's no obvious sentinel value that can be used for "this item
was defaulted", at least not without a lookup table.

It should be sufficient for a long time - should a more complete
representation be directly generatble later then that could be done
rather invisibly for downstream clients.

This approach leads to some simplification of the basic rules that parse
record entries, which have been blessed with slightly more thorough
testing. They're also easier to compose, since they don't have to
enforce element-wise type checking, however it's possible to require
all-int, all-float, all-string which will be more strictly type checked,
or a mixed int-string or int-float keyword that enjoy slightly less
strict checking (i.e. it doesn't guard against string entries in int
positions).

This introduces the small umbra program that reads a deck and outputs
its corresponding graphviz representation, which is useful for debugging
and visualisation. Eventually this will be backed by a proper library
with some more bells and whistles.