Add concatenate
Dealing with includes is quite horrible, and boost-spirit isn't too well
equipped to deal with parsing files that requires pausing and resuming
with a new file.

The concatenate function does a scan of the input and look for all
INCLUDE and PATH keywords, and copies the rest into a vector. This
vector is then later presented to the parser, which now does not have to
deal with the include logic. This comes at a memory cost linear with
input files, but is rather fast and reduces a lot of complexity in the
parser logic.

This could've been done with a custom iterator that does this on the
fly, but since it has to be bidirectional and cheap to copy that quickly
becomes complicated and possibly slow.

Doing the concatenation also opens up for a small program that *just*
performs all the includes, which can be useful sometimes.