Dynamically (lazy) look up the parser from keyword
Since the grammar is increadibly keyword -> records oriented, and the
spirit parsers rely on backtracking, it makes sense to select the parser
run-time based on the keyword. Luckily, the Nabialek trick exists.

Merge the grammar objects to one and use the Nabialek trick internally.

This should enable easier run-time extension (or even redefinition) of
keywords as well.