Represent the AST with less nested data structures
Redesign the first parsing pass to not be section aware, and remove the
explicit representation of records-as-vectors, and instead represent all
keyword data as a single vector, where records are separated by special
endrec types.

The goal for this change is to create a parser that can be used on
arbitrary incomplete decks that will still build an AST, and do all
semantic checks on that. While section -> set-of-productions could be
considered a reasonable view of the grammar, treating section
introductions as toggle keywords opens up for more capabilities of
parsing arbitrary sub input with the mother parser, without having to
expose multiple less-complete parsers.

Down the line, this might give some performance gains as traversing
chunky keywords is less pointer jumping, and less allocations have to be
done.