Add pseudo-stable AST interface
To reduce some time playing catch-up with development, the ast+cursor
interface in ast.hpp is introduced as a semi-stable target for
applications and other parts of the library to use while things are in
flux. Other parts of the lib can target this, and changes in the grammar
module will only leak into ast.cpp

This interface is in no way the final, stable, user-facing API to the
rest of the library, nor end-users. It does limit the surface of boost,
a much desired property, and should make most other development faster.